+++ README.rst
@@ -56,22 +56,15 @@ Usage
 
    .. code-block:: bash
 
-    # This cookiecutter provides support for initializing the database either
-    # (a) with migrations through alembic, or
-    # (b) without migrations.
+    # This cookiecutter provides support for initializing the database
+    # with migrations through alembic.
     #
-    # (a) Initialize the database with migrations through alembic:
+    # Initialize the database with migrations through alembic:
     # First use alembic to generate revisions.
     $ env/bin/alembic -c development.ini revision --autogenerate -m "init"
     # upgrade to that revision:
     $ env/bin/alembic -c development.ini upgrade head
     # Later you can run database migrations.
-    # Also there is commented code that can be used to
-    # initialize the database directly to the latest revision.
-    # {{cookiecutter.repo_name}}/scripts/initializedb.py
-    #
-    # (b) Initialize the database without migrations:
-    $ env/bin/initialize_{{cookiecutter.repo_name}}_db development.ini
 
 4. Run your project's tests.
