+++ hooks/post_gen_project.py
@@ -48,22 +48,15 @@ msg = dedent(
 
     Initialize the database.
 
-        # This cookiecutter provides support for initializing the database either
-        # (a) with migrations through alembic, or
-        # (b) without migrations.
+        # This cookiecutter provides support for initializing the database
+        # with migrations through alembic.
         #
-        # (a) Initialize the database with migrations through alembic:
+        # Initialize the database with migrations through alembic:
         # First use alembic to generate revisions.
         %(alembic_cmd)s -c development.ini revision --autogenerate -m "init"
         # upgrade to that revision:
         %(alembic_cmd)s -c development.ini upgrade head
         # Later you can run database migrations.
-        # Also there is commented code that can be used to
-        # initialize the database directly to the latest revision.
-        # {{cookiecutter.repo_name}}/scripts/initializedb.py
-        #
-        # (b) Initialize the database without migrations:
-        %(init_cmd)s development.ini
 
     Run your project's tests.
         %(pytest_cmd)s
