+++ {{cookiecutter.repo_name}}/development.ini
@@ -29,7 +29,7 @@ retry.attempts = 3
 [alembic]
 # path to migration scripts
 script_location = {{ cookiecutter.repo_name }}/alembic
-file_template = %%(year)d_%%(month).2d_%%(day).2d_%%(hour).2d_%%(minute).2d_%%(second).2d_%%(rev)s
+file_template = %%(year)d%%(month).2d%%(day).2d_%%(rev)s
 # file_template = %%(rev)s_%%(slug)s
 
 [server:main]
