--- {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/scripts/initializedb.py
@@ -1,47 +1,0 @@
-import os
-import sys
-
-from pyramid.paster import get_appsettings, setup_logging
-from pyramid.scripts.common import parse_vars
-
-import transaction
-
-from ..models import MyModel, get_engine, get_session_factory, get_tm_session
-from ..models.meta import Base
-
-
-# from alembic import command
-# from alembic.config import Config
-
-
-def usage(argv):
-    cmd = os.path.basename(argv[0])
-    print('usage: %s <config_uri> [var=value]\n'
-          '(example: "%s development.ini")' % (cmd, cmd))
-    sys.exit(1)
-
-
-def main(argv=sys.argv):
-    if len(argv) < 2:
-        usage(argv)
-    config_uri = argv[1]
-    options = parse_vars(argv[2:])
-    setup_logging(config_uri)
-    settings = get_appsettings(config_uri, options=options)
-
-    engine = get_engine(settings)
-    Base.metadata.create_all(engine)
-
-    # http://alembic.zzzcomputing.com/en/latest/cookbook.html#building-an-up-to-date-database-from-scratch
-    # Load the Alembic configuration and generate the
-    # version table, "stamping" it with the most recent rev:
-    # alembic_cfg = Config(config_uri)
-    # command.stamp(alembic_cfg, "head")
-
-    session_factory = get_session_factory(engine)
-
-    with transaction.manager:
-        dbsession = get_tm_session(session_factory, transaction.manager)
-
-        model = MyModel(name='one', value=1)
-        dbsession.add(model)
