+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/templates/mytemplate.jinja2
