+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/views/default.py
@@ -20,9 +20,8 @@ db_err_msg = """\
 Pyramid is having a problem using your SQL database.  The problem
 might be caused by one of the following things:
 
-1.  You may need to run the "initialize_{{ cookiecutter.repo_name }}_db" script
-    to initialize your database tables.  Check your virtual
-    environment's "bin" directory for this script and try to run it.
+1.  You may need to initialize your database tables with `alembic`.
+    Check your README.txt for description and try to run it.
 
 2.  Your database server may not be running.  Check that the
     database server referred to by the "sqlalchemy.url" setting in
